[fix] 생성자 기반 회원가입 로직 재구성
